# optimize async log